Remove switch statement from environment selection
Seemed a bit fragile to have to keep numbers in a switch statement in
line with the text presented to the user. This uses an array of structs
named `ENVIRONMENTS` so environments and their corresponding baseurl
can be defined and updated in a single place without needing to update
any of the underlying methods.